# go/APIvidgm/docs/swagger.yaml
definitions:
...
  models.Cart:
    properties:
      id:
        type: integer
      total_price:
        type: number
      user_id:
        type: integer
      videogames:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
    type: object
  models.CartItem:
    properties:
      cart_id:
        type: integer
      id:
        type: integer
      is_physical:
        type: boolean
      quantity:
        type: integer
      videogame:
        $ref: '#/definitions/models.VideoGame'
      videogame_id:
        type: integer
    type: object
...
  models.Genre:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Label:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Platform:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Review:
    properties:
      comment:
        type: string
      id:
        type: integer
      modstatus:
        description: pending, approved, rejected
        type: string
      rating:
        type: integer
      user_id:
        type: integer
      username:
        type: string
      videogame_id:
        type: integer
      videogame_name:
        type: string
    type: object
...
  models.VideoGame:
    properties:
      developer:
        type: string
      dig_stock:
        type: integer
      genre:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: integer
      image_url:
        type: string
      label:
        items:
          $ref: '#/definitions/models.Label'
        type: array
      min_req:
        type: string
      pegi:
        type: integer
      phy_stock:
        type: integer
      platform:
        items:
          $ref: '#/definitions/models.Platform'
        type: array
      price:
        type: number
      rating:
        type: number
      release_year:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      sum_rating:
        type: integer
      synopsis:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Shit documentation for shit API.
  termsOfService: http://swagger.io/terms/
  title: VideoGames API
  version: "1.0"
paths:
  /cart/{id}/checkout:
    post:
      consumes:
      - application/json
      description: Processes the checkout of a user's cart, creating digital and/or
        physical orders. Physical orders require a delivery address ID provided via
        query parameter `dirID`. Requires authentication and the same user ID.
      parameters:
      - description: Cart ID
        in: path
        name: id
        required: true
        type: integer
      - description: Delivery address ID (required if cart contains physical items)
        in: query
        name: dirID
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Checkout completed successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid or missing parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Cart not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Checkout user's cart
      tags:
      - Cart
...
  /users/{id}/cart:
    get:
      consumes:
      - application/json
      description: Retrieves a user's cart along with its video games. If the cart
        does not exist, a new one is created automatically. Requires authentication
        and the same user ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Cart'
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user's cart by user ID
      tags:
      - Cart
  /users/{id}/cart/items:
    post:
      consumes:
      - application/json
      description: Adds a video game to the user's cart. If the cart does not exist,
        it is created automatically. Requires authentication and the same user ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cart item data (includes VideoGameID and Quantity)
        in: body
        name: cartItem
        required: true
        schema:
          $ref: '#/definitions/models.CartItem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CartItem'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User or video game not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add an item to a user's cart
      tags:
      - Cart
  /users/{id}/cart/items/{itemid}:
    delete:
      consumes:
      - application/json
      description: Removes a specific video game from the user's cart and updates
        the total price. Requires authentication and the same user ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cart item ID
        in: path
        name: itemid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cart item deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User, cart, or cart item not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a cart item
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Updates the quantity of a specific cart item for a user. Recalculates
        the cart's total price. Requires authentication and the same user ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cart item ID
        in: path
        name: itemid
        required: true
        type: integer
      - description: Updated cart item data (quantity)
        in: body
        name: cartItem
        required: true
        schema:
          $ref: '#/definitions/models.CartItem'
      produces:
      - application/json
      responses:
        "200":
          description: Updated total price
          schema:
            additionalProperties:
              format: float64
              type: number
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User, cart, or cart item not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a cart item
      tags:
      - Cart
...
swagger: "2.0"